feat: remove jenkins-github-integ plugin
