--- docs/plugins/jenkins-github-integ.md
@@ -1,8 +1,0 @@
-# jenkins-github-integ plugin
-
-TODO(aFlyBird0): Add your document here.
-## Usage
-
-```yaml
---8<-- "jenkins-github-integ.yaml"
-```
