--- internal/pkg/plugin/jenkinsgithub/jenkinsgithub_suite_test.go
@@ -1,13 +1,0 @@
-package jenkinsgithub
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestJenkinsgithub(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Jenkinsgithub Suite")
-}
