--- internal/pkg/plugin/jenkinsgithub/tpl/github-integ.tpl.yaml
@@ -1,23 +1,0 @@
-unclassified:
-  ghprbTrigger:
-    adminlist: "[[range .AdminList]][[.]] [[end]]"
-    autoCloseFailedPullRequests: false
-    cron: "H/5 * * * *"
-    extensions:
-    - ghprbSimpleStatus:
-        addTestResults: false
-        showMatrixStatus: false
-    githubAuth:
-    - credentialsId: "[[.CredentialsID]]"
-      description: "Anonymous connection"
-      id: "[[.GithubAuthID]]"
-      jenkinsUrl: "[[.JenkinsURLOverride]]"
-      serverAPIUrl: "https://api.github.com"
-    manageWebhooks: true
-    okToTestPhrase: ".*ok\\W+to\\W+test.*"
-    requestForTestingPhrase: "Can one of the admins verify this patch?"
-    retestPhrase: ".*test\\W+this\\W+please.*"
-    skipBuildPhrase: ".*\\[skip\\W+ci\\].*"
-    useComments: false
-    useDetailedComments: false
-    whitelistPhrase: ".*add\\W+to\\W+whitelist.*"
