--- internal/pkg/plugin/jenkinsgithub/tpl/job-pr.tpl.xml
@@ -1,82 +1,0 @@
-<?xml version='1.1' encoding='UTF-8'?>
-<flow-definition plugin="workflow-job@1207.ve6191ff089f8">
-    <description>A pipeline is only triggered when pr.</description>
-    <keepDependencies>false</keepDependencies>
-    <properties>
-        <com.coravy.hudson.plugins.github.GithubProjectProperty plugin="github@1.34.4">
-            <projectUrl>[[.GitHubRepoURL]]</projectUrl>
-            <displayName></displayName>
-        </com.coravy.hudson.plugins.github.GithubProjectProperty>
-        <org.jenkinsci.plugins.workflow.job.properties.PipelineTriggersJobProperty>
-            <triggers>
-                <org.jenkinsci.plugins.ghprb.GhprbTrigger plugin="ghprb@1.42.2">
-                    <spec>H/5 * * * *</spec>
-                    <configVersion>3</configVersion>
-                    <adminlist>[[range .AdminList]][[.]] [[end]]</adminlist>
-                    <allowMembersOfWhitelistedOrgsAsAdmin>false</allowMembersOfWhitelistedOrgsAsAdmin>
-                    <orgslist></orgslist>
-                    <cron>H/5 * * * *</cron>
-                    <buildDescTemplate></buildDescTemplate>
-                    <onlyTriggerPhrase>false</onlyTriggerPhrase>
-                    <useGitHubHooks>true</useGitHubHooks>
-                    <permitAll>false</permitAll>
-                    <whitelist></whitelist>
-                    <autoCloseFailedPullRequests>false</autoCloseFailedPullRequests>
-                    <displayBuildErrorsOnDownstreamBuilds>false</displayBuildErrorsOnDownstreamBuilds>
-                    <whiteListTargetBranches>
-                        <org.jenkinsci.plugins.ghprb.GhprbBranch>
-                            <branch></branch>
-                        </org.jenkinsci.plugins.ghprb.GhprbBranch>
-                    </whiteListTargetBranches>
-                    <blackListTargetBranches>
-                        <org.jenkinsci.plugins.ghprb.GhprbBranch>
-                            <branch></branch>
-                        </org.jenkinsci.plugins.ghprb.GhprbBranch>
-                    </blackListTargetBranches>
-                    <gitHubAuthId>3a3b9ece-ad38-4209-8808-a37fbe74cc95</gitHubAuthId>
-                    <triggerPhrase></triggerPhrase>
-                    <skipBuildPhrase>.*\[skip\W+ci\].*</skipBuildPhrase>
-                    <blackListCommitAuthor></blackListCommitAuthor>
-                    <blackListLabels></blackListLabels>
-                    <whiteListLabels></whiteListLabels>
-                    <includedRegions></includedRegions>
-                    <excludedRegions></excludedRegions>
-                    <extensions>
-                        <org.jenkinsci.plugins.ghprb.extensions.status.GhprbSimpleStatus>
-                            <commitStatusContext></commitStatusContext>
-                            <triggeredStatus></triggeredStatus>
-                            <startedStatus></startedStatus>
-                            <statusUrl></statusUrl>
-                            <addTestResults>false</addTestResults>
-                        </org.jenkinsci.plugins.ghprb.extensions.status.GhprbSimpleStatus>
-                    </extensions>
-                </org.jenkinsci.plugins.ghprb.GhprbTrigger>
-            </triggers>
-        </org.jenkinsci.plugins.workflow.job.properties.PipelineTriggersJobProperty>
-    </properties>
-    <definition class="org.jenkinsci.plugins.workflow.cps.CpsScmFlowDefinition" plugin="workflow-cps@2759.v87459c4eea_ca_">
-        <scm class="hudson.plugins.git.GitSCM" plugin="git@4.11.3">
-            <configVersion>2</configVersion>
-            <userRemoteConfigs>
-                <hudson.plugins.git.UserRemoteConfig>
-                    <name>origin</name>
-                    <refspec>+refs/pull/*:refs/remotes/origin/pr/*</refspec>
-                    <url>[[.GitHubRepoURL]]</url>
-                    <credentialsId>[[.CredentialsID]]</credentialsId>
-                </hudson.plugins.git.UserRemoteConfig>
-            </userRemoteConfigs>
-            <branches>
-                <hudson.plugins.git.BranchSpec>
-                    <name>${sha1}</name>
-                </hudson.plugins.git.BranchSpec>
-            </branches>
-            <doGenerateSubmoduleConfigurations>false</doGenerateSubmoduleConfigurations>
-            <submoduleCfg class="empty-list"/>
-            <extensions/>
-        </scm>
-        <scriptPath>[[.PrPipelineScriptPath]]</scriptPath>
-        <lightweight>false</lightweight>
-    </definition>
-    <triggers/>
-    <disabled>false</disabled>
-</flow-definition>
