--- internal/pkg/plugin/jenkinsgithub/tpl/safe-html.tpl.yaml
@@ -1,4 +1,0 @@
-jenkins:
-  markupFormatter:
-    rawHtml:
-      disableSyntaxHighlighting: false
