--- internal/pkg/show/config/plugins/jenkins-github-integ.yaml
@@ -1,35 +1,0 @@
-tools:
-# name of the tool
-- name: jenkins-github-integ
-  # id of the tool instance
-  instanceID: default
-  # format: name.instanceID; If specified, dtm will make sure the dependency is applied first before handling this tool.
-  dependsOn: [ ]
-  # options for the plugin
-  options:
-    jenkins:
-      # jenkinsUrl, format: hostname:port
-      url: localhost:8080
-      # override the jenkins url to expose the jenkins to the GitHub webhook, can be empty.
-      urlOverride:
-      # jenkins user name, default: admin
-      user: admin
-      # jenkins password, you have 3 options to set the password:
-      # 1. use outputs of the `jenkins` plugin, see docs for more details
-      # 2. set the `JENKINS_PASSWORD` environment variable
-      # 3. fill in the password in this field(not recommended)
-      # if all set, devstream will read the password from the config file first.
-      password:
-      # jenkins job name, mandatory
-      jobName:
-      # path to the pipeline file, relative to the git repo root directory. default: Jenkinsfile-pr
-      pipelineScriptPath: Jenkinsfile-pr
-    helm:
-      # namespace of the jenkins, default: jenkins
-      namespace: jenkins
-      # release name of the jenkins helm chart, mandatory
-      releaseName:
-    # GitHub repo where to put the pipeline script and project. mandatory
-    githubRepoUrl: https://github.com/YOUR_GITHUB_ACCOUNT/YOUR_TEST_PROJECT_NAME
-    # list of GitHub admin users
-    adminList: []
